requirements.txt
aws-cdk.aws-apigateway
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudwatch
aws-cdk.aws-dynamodb
aws-cdk.aws-iam
aws-cdk.aws-lambda
aws-cdk.aws-s3
aws-cdk.aws-s3-deployment
aws-cdk.aws-sagemaker
aws-cdk.aws-iam
aws_cdk.aws_ec2
aws-cdk.aws_cloud9
aws-cdk.aws_codecommit
aws-cdk.aws-glue
aws-cdk.aws-lambda-event-sources
arrow==0.15.5
astroid==1.6.6
attrs==20.3.0
aws-lambda-builders==0.6.0
aws-sam-cli==0.38.0
aws-sam-translator==1.19.0
awscli
backcall==0.2.0
binaryornot==0.4.4
blis==0.4.1
boto3
botocore
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
chevron==0.13.1
click==7.1.2
colorama==0.4.3
cookiecutter==1.6.0
cymem==2.0.3
dateparser==0.7.2
datetime==4.3
decorator==4.4.2
Django==2.0.2
docker==4.2.0
docutils==0.15.2
Flask==1.0.4
future==0.18.2
geonamescache==1.2.0
git-remote-codecommit==1.13
idna==2.8
ikp3db==1.4.1
importlib-metadata==1.7.0
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.1.1
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.1
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==3.2.0
langdetect==1.0.8
lazy-object-proxy==1.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.0
parso==0.1.1
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
pip==9.0.3
plac==1.1.3
poyo==0.5.0
preshed==3.0.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.6.1
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.5
s3transfer==0.3.3
serverlessrepo==0.1.9
setuptools==36.2.7
six==1.15.0
spacy==2.3.2
srsly==1.0.2
stevedore==3.1.0
thinc==7.4.1
tomlkit==0.5.8
tqdm==4.48.0
traitlets==4.3.3
twython==3.8.2
tzlocal==2.0.0
urllib3==1.25.9
virtualenv==16.2.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wasabi==0.7.1
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.0
wheel==0.34.2
whichcraft==0.6.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0